updated class according to review
I looked at all of your comments and tried fixing the solution according to them (hope I didn't miss any).
The main file isn't really different because I understand someone else will rewrite the tests, but you can look in it to see how the class is used.
It's simple (I think): initialize an instance with a dictionary (or set a dictionary with setDictionary() to an existing instance), use getAlphabet() to get the alphabet, use isDictionaryConsistent() to find if the dictionary is consistent.
Cycles in the graph are found during the creation of the graph, I think it's efficient since I add an O(1) action to each iteration, but I'm happy to hear your opinion about it.
Another thing I did and I wanted to ask your opinion about - when I find that the dictionary is inconsistent, I set the alphabet to be the characters of the dictionary in an arbitrary order. In the previous solution, I returned an empty alphabet. What do you think?